+++ Code/CrowdControlMod.cs
@@ -926,6 +926,9 @@ public sealed class CrowdControlMod : Mod
         AddEffect(new MoneyBoostEffect(60));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Pickaxe));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Sword));
+        AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Magic));
+        AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Summon));
+        AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Ranged));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Armour));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.HealingPotion));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Potion));